+++ src/ffpack/lcc/johannessonCounting.py
@@ -9,7 +9,7 @@ Probability in the Engineering and Informational Sciences, 12(2), pp.143-175.
 '''
 
 import numpy as np
-from ffpack.utils import generalUtils 
+from ffpack.utils import sequenceFilter 
 from ffpack.config import globalConfig
 from collections import defaultdict 
 
@@ -66,7 +66,7 @@ def johannessonMinMaxCounting( data, aggregate=True ):
 
     # we need to use this util function since it keeps one peak 
     # if there are two or more points together with the same peak value
-    data = generalUtils.sequencePeakValleyFilter( data, keepEnds=True )
+    data = sequenceFilter.sequencePeakValleyFilter( data, keepEnds=True )
     rstSeq = [ ]
     for i in range( 1, len( data ) - 1 ):
         if ( data[ i ] > data[ i - 1 ] and data[ i ] > data[ i + 1 ] ):
