+++ src/ffpack/utils/__init__.py
@@ -1,5 +1,6 @@
-from .generalUtils import *
-from .fdrUtils import *
-from .lccUtils import *
-from .lsmUtils import *
-from .miscUtils import *
+from .aggregation import *
+from .countingMatrix import *
+from .derivatives import *
+from .digitization import *
+from .fitter import *
+from .sequenceFilter import *
