+++ src/ffpack/utils/countingMatrix.py
