+++ src/ffpack/utils/fitter.py
