+++ tests/lcc/test_lcc_astmCounting.py
@@ -285,7 +285,6 @@ def test_astmLevelCrossingCounting_withLevelsCase_pass():
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-
 ###############################################################################
 # Test astmPeakCounting function
 ###############################################################################
@@ -470,7 +469,6 @@ def test_astmPeakCounting_withRefLevelCase_pass():
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-
 ###############################################################################
 # Test astmSimpleRangeCounting function
 ###############################################################################
@@ -495,7 +493,7 @@ def test_astmSimpleRangeCounting_twoDimInputCase_valueError():
         _ = lcc.astmSimpleRangeCounting( data )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmSimpleRangeCounting_normalUseCase_pass( mock_get ):
     # Standard simple range counting data from E1049-85(2017) Fig.4(a)
     data = [ -2.0, 1.0, -3.0, 5.0, -1.0, 3.0, -4.0, 4.0, -2.0 ]
@@ -527,7 +525,7 @@ def test_astmSimpleRangeCounting_normalUseCase_pass( mock_get ):
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmSimpleRangeCounting_normalTrivialCase_pass( mock_get ):
     # Trivial case 1
     data = [ 0.0, 1.5 ]
@@ -653,7 +651,6 @@ def test_astmSimpleRangeCounting_normalTrivialCase_pass( mock_get ):
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-
 ###############################################################################
 # Test astmRainflowCounting function
 ###############################################################################
@@ -678,7 +675,7 @@ def test_astmRainflowCounting_twoDimInputCase_valueError():
         _ = lcc.astmRainflowCounting( data )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmRainflowCounting_normalUseCase_pass( mock_get ):
     # Standard rainflow counting data from E1049-85(2017) Fig.6(a)
     data = [ -2.0, 1.0, -3.0, 5.0, -1.0, 3.0, -4.0, 4.0, -2.0 ]
@@ -708,7 +705,7 @@ def test_astmRainflowCounting_normalUseCase_pass( mock_get ):
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmRainflowCounting_normalTrivialCase_pass( mock_get ):
     # Trivial case 1
     data = [ 0.0, 1.5 ]
@@ -832,7 +829,6 @@ def test_astmRainflowCounting_normalTrivialCase_pass( mock_get ):
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-
 ###############################################################################
 # Test astmRangePairCounting function
 ###############################################################################
@@ -854,7 +850,7 @@ def test_astmRangePairCounting_twoDimInputCase_valueError():
         _ = lcc.astmRangePairCounting( data )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmRangePairCounting_twoPoints_empty( mock_get ):
     data = [ -2.0, 1.0 ]
     mock_get.return_value = [ -2.0, 1.0 ]
@@ -867,7 +863,7 @@ def test_astmRangePairCounting_twoPoints_empty( mock_get ):
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmRangePairCounting_threePoints_oneCount( mock_get ):
     # case 1: same left and right height
     data = [ -2.0, 1.0, -2.0 ]
@@ -903,7 +899,7 @@ def test_astmRangePairCounting_threePoints_oneCount( mock_get ):
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmRangePairCounting_fourPoints_oneCount( mock_get ):
     # case 1: same left and right height
     data = [ -2.0, 1.0, -2.0, 1.0 ]
@@ -939,7 +935,7 @@ def test_astmRangePairCounting_fourPoints_oneCount( mock_get ):
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmRangePairCounting_fivePoints_depends( mock_get ):
     # case 1: same left and right height
     data = [ -2.0, 1.0, -2.0, 1.0, -2.0 ]
@@ -986,7 +982,7 @@ def test_astmRangePairCounting_fivePoints_depends( mock_get ):
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmRangePairCounting_normalUseCase_pass( mock_get ):
     # range pair counting data from E1049-85(2017) Fig.5
     data = [ -2.0, 1.0, -3.0, 5.0, -1.0, 3.0, -4.0, 4.0, -2.0 ]
@@ -1000,7 +996,7 @@ def test_astmRangePairCounting_normalUseCase_pass( mock_get ):
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmRangePairCounting_duplicatedHeight_aggregated( mock_get ):
     # modified range pair counting data from E1049-85(2017) Fig.5
     data = [ -2.0, 1.0, -2.0, 5.0, -1.0, 2.0, -4.0, 4.0, -2.0 ]
@@ -1012,7 +1008,6 @@ def test_astmRangePairCounting_duplicatedHeight_aggregated( mock_get ):
     calRst = lcc.astmRangePairCounting( data, aggregate=True )
     expectedRst = [ [ 3.0, 2 ], [ 6.0, 1 ], [ 7.0, 1 ] ]
     np.testing.assert_allclose( calRst, expectedRst )
-
 
 
 ###############################################################################
@@ -1042,7 +1037,7 @@ def test_astmRainflowRepeatHistoryCounting_dataNotRepeating_valueError():
         _ = lcc.astmRainflowRepeatHistoryCounting( data )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmRainflowRepeatHistoryCounting_twoDataPoint_empty( mock_get ):
     data = [ 1.0, 1.0 ]
     mock_get.return_value = [ 1.0, 1.0 ]
@@ -1055,7 +1050,7 @@ def test_astmRainflowRepeatHistoryCounting_twoDataPoint_empty( mock_get ):
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmRainflowRepeatHistoryCounting_threeDataPoint_oneCount( mock_get ):
     data = [ 1.0, 2.0, 1.0 ]
     peakValley1 = [ 1.0, 2.0, 1.0 ]
@@ -1079,7 +1074,7 @@ def test_astmRainflowRepeatHistoryCounting_threeDataPoint_oneCount( mock_get ):
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmRainflowRepeatHistoryCounting_fourDataPoint_depends( mock_get ):
     # case 1: increase first - one point in the middle
     data = [ 1.0, 2.0, -1.0, 1.0 ]
@@ -1120,7 +1115,7 @@ def test_astmRainflowRepeatHistoryCounting_fourDataPoint_depends( mock_get ):
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmRainflowRepeatHistoryCounting_fiveDataPoint_depends( mock_get ):
     # case 1: w final - different heights
     data = [ 1.0, 2.0, -1.0, 2.0, 1.0 ]
@@ -1159,7 +1154,7 @@ def test_astmRainflowRepeatHistoryCounting_fiveDataPoint_depends( mock_get ):
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-@patch( "ffpack.utils.generalUtils.sequencePeakValleyFilter" )
+@patch( "ffpack.utils.sequenceFilter.sequencePeakValleyFilter" )
 def test_astmRainflowRepeatHistoryCounting_normalUseCase_pass( mock_get ):
     # simplified rainflow counting data for repeating histories from E1049-85(2017) Fig.7
     data = [ -2.0, 1.0, -3.0, 5.0, -1.0, 3.0, -4.0, 4.0, -2.0 ]
