+++ tests/utils/test_utils_countingMatrix.py
@@ -4,6 +4,7 @@ from ffpack import utils
 import numpy as np
 import pytest
 
+
 ###############################################################################
 # Test countingRstToCountingMatrix
 ###############################################################################
